SAROJ-2 Split functionality in 2 files
